+++ src/app/code/community/Shopgate/Cloudapi/Model/Frontend/Observer/OnepageSuccessAction.php
@@ -24,8 +24,6 @@ class Shopgate_Cloudapi_Model_Frontend_Observer_OnepageSuccessAction
     /**
      * Checks if the order is received from Shopgate API call.
      *
-     * @todo-sg: send the order to the pipeline for tracking
-     *
      * @param Varien_Event_Observer $observer
      */
     public function execute(Varien_Event_Observer $observer)
@@ -33,6 +31,25 @@ class Shopgate_Cloudapi_Model_Frontend_Observer_OnepageSuccessAction
         $session = Mage::getSingleton('checkout/session');
 
         if ($session->getData(Shopgate_Cloudapi_Helper_Frontend_Checkout::SESSION_IS_SHOPGATE_CHECKOUT)) {
+            $orderIds = $observer->getEvent()->getOrderIds();
+            if (isset($orderIds[0])) {
+                $newOrderId = $orderIds[0];
+                $layout     = Mage::app()->getLayout();
+                /** @var Shopgate_Cloudapi_Block_PipelineRequest $pipelineRequestBlock */
+                $pipelineRequestBlock = $layout->createBlock('shopgate_cloudapi/pipelineRequest');
+                $pipelineRequestBlock->addMethod(
+                    array(
+                        'serial' => Shopgate_Cloudapi_Block_PipelineRequest::PIPELINE_REQUEST_SERIAL, //@todo evaluate the right serial
+                        'name'   => Shopgate_Cloudapi_Block_PipelineRequest::PIPELINE_REQUEST_CREATE_NEW_CART_FOR_CUSTOMER,
+                        'input'  => array(
+                            'orderId' => $newOrderId
+                        )
+                    )
+                );
+                $pipelineRequestBlock->setTemplate('shopgate/cloudapi/pipelineRequest.phtml');
+                $layout->getBlock('head')->addJs('shopgate/pipelineRequest.js');
+                $layout->getBlock('head')->append($pipelineRequestBlock);
+            }
             $session->unsetData(Shopgate_Cloudapi_Helper_Frontend_Checkout::SESSION_IS_SHOPGATE_CHECKOUT);
         }
     }