+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## [Unreleased]
 ### Added
 - Add shopgate_order_sources Table to mark order as Shopgate order
+- Add utm parameters for Google Analytics in Web checkout
 
 ## 3.0.0 - 2017-11-23
 ### Added