+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- shopgate_order_sources table to flag Shopgate orders
+- UTM parameters for Google Analytics in web checkout
 
 ## 3.0.0 - 2017-11-23
 ### Added