+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - shopgate_order_sources table to flag Shopgate orders
 - UTM parameters for Google Analytics in web checkout
 - Endpoint for retrieving basic customer information
+- Added logic for custom layout handles and removed unnecessary elements from app webview
 
 ## 3.0.0 - 2017-11-23
 ### Added