+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## 3.1.0 - 2018-03-15
 ### Added
 - shopgate_order_sources table to flag Shopgate orders
 - UTM parameters for Google Analytics in web checkout